Encapsulate arena atmosphere per GLB definition
Introduce arenas/ module with registry pattern: each arena pairs a glbUrl
with its own Atmosphere component (fog, HDRI, lighting). Desert arena owns
its quarry HDRI and warm lighting in arenas/desert/. ArenaEnvironment
selects definition by arenaId for future multi-biome support.
